// frontend/src/app/model/search.model.ts
export class Search {
    userId: number;
    numOfPersons: number;
    name: string;
    location: number;
    music: string;
    accomodation: string;
    smokingArea: boolean;
    nonSmokingArea: boolean;
    alcoholicDrinks: boolean;
    nonAlcoholicDrinks: boolean;
    petFriendly: boolean;
    kidFriendly: boolean;
    rate: number;
}